Also expose generateThemeStylesheet to theming-build.
