+++ src/browser/index.ts
@@ -1,38 +1,10 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { ThemePreset, Override, validateOverride } from '../shared/theme';
-import { createOverrideDeclarations, createScopedThemeDeclarations } from '../shared/declaration';
+import { Override, ThemePreset } from '../shared/theme';
+import { generateThemeStylesheet } from '../shared/theme-stylesheet';
 import { getNonce, createStyleNode, appendStyleNode } from './dom';
-import { createMultiThemeCustomizer } from '../shared/declaration/customizer';
-import { getContexts, getThemeFromPreset, validateScopedThemeOverride } from '../shared/theme/validate';
 
-export interface GenerateThemeStylesheetParams {
-  override: Override;
-  preset: ThemePreset;
-  baseThemeId?: string;
-
-  selector?: string;
-}
-
-export function generateThemeStylesheet(params: GenerateThemeStylesheetParams): string {
-  const { override, preset, baseThemeId, selector } = params;
-  const availableContexts = getContexts(preset);
-  const validated = validateOverride(override, preset.themeable, availableContexts);
-  const theme = getThemeFromPreset(preset, baseThemeId);
-
-  const scopedSelector = selector?.trim();
-  if (scopedSelector) {
-    validateScopedThemeOverride(theme, validated);
-    return createScopedThemeDeclarations(theme, validated, preset.propertiesMap, scopedSelector);
-  }
-
-  return createOverrideDeclarations(
-    theme,
-    validated,
-    preset.propertiesMap,
-    createMultiThemeCustomizer(preset.theme.selector),
-  );
-}
+export { generateThemeStylesheet, GenerateThemeStylesheetParams } from '../shared/theme-stylesheet';
 
 export interface ApplyThemeParams {
   override: Override;
